refactor: move entities table into re-usable component
