+++ apps/hash-frontend/src/pages/[shortname]/types/entity-type/[...slug-maybe-version].page/entities-tab.tsx
@@ -3,51 +3,28 @@ import {
   faAsterisk,
   faMagnifyingGlass,
 } from "@fortawesome/free-solid-svg-icons";
-import { GridCellKind, Item, TextCell } from "@glideapps/glide-data-grid";
 import { Chip, FontAwesomeIcon, IconButton } from "@hashintel/design-system";
 import { extractOwnedByIdFromEntityId } from "@local/hash-subgraph";
 import { Box, Paper, Stack } from "@mui/material";
-import { useRouter } from "next/router";
-import {
-  FunctionComponent,
-  useCallback,
-  useContext,
-  useMemo,
-  useState,
-} from "react";
+import { FunctionComponent, useContext, useMemo, useState } from "react";
 
-import { Grid } from "../../../../../components/grid/grid";
-import { BlankCell, blankCell } from "../../../../../components/grid/utils";
 import { EarthAmericasRegularIcon } from "../../../../../shared/icons/earth-americas-regular";
 import { HomeIcon } from "../../../../../shared/icons/home-icon";
+import { EntitiesTable } from "../../../../shared/entities-table";
 import { WorkspaceContext } from "../../../../shared/workspace-context";
 import { SectionEmptyState } from "../../../shared/section-empty-state";
 import { SectionWrapper } from "../../../shared/section-wrapper";
 import { WhiteChip } from "../../../shared/white-chip";
-import {
-  renderTextIconCell,
-  TextIconCell,
-} from "./entities-tab/text-icon-cell";
-import {
-  TypeEntitiesRow,
-  useEntitiesTable,
-} from "./entities-tab/use-entities-table";
 import { useEntityType } from "./shared/entity-type-context";
 import { useEntityTypeEntities } from "./shared/entity-type-entities-context";
 
 export const EntitiesTab: FunctionComponent = () => {
-  const router = useRouter();
+  const [_showSearch, setShowSearch] = useState<boolean>(false);
 
-  const { entities, entityTypes, propertyTypes, subgraph } =
-    useEntityTypeEntities();
-
-  const [showSearch, setShowSearch] = useState(false);
+  const { entities } = useEntityTypeEntities();
 
   const { activeWorkspaceAccountId, activeWorkspace } =
     useContext(WorkspaceContext);
-
-  const { columns, rows } =
-    useEntitiesTable(entities, entityTypes, propertyTypes, subgraph) ?? {};
 
   const entityType = useEntityType();
 
@@ -64,51 +41,6 @@ export const EntitiesTab: FunctionComponent = () => {
       public: (entities?.length ?? 0) - namespaceEntities.length,
     };
   }, [entities, activeWorkspaceAccountId]);
-
-  const createGetCellContent = useCallback(
-    (entityRows: TypeEntitiesRow[]) =>
-      ([colIndex, rowIndex]: Item): TextIconCell | TextCell | BlankCell => {
-        if (columns) {
-          const row = entityRows[rowIndex];
-          const columnId = columns[colIndex]?.id;
-          const cellValue = columnId && row?.[columnId];
-
-          if (cellValue) {
-            if (columnId === "entity") {
-              return {
-                kind: GridCellKind.Custom,
-                allowOverlay: false,
-                readonly: true,
-                copyData: cellValue,
-                cursor: "pointer",
-                data: {
-                  kind: "text-icon-cell",
-                  icon: "bpAsterisk",
-                  value: cellValue,
-                  onClick: () =>
-                    router.push(`/${row.namespace}/entities/${row.entityId}`),
-                },
-              };
-            }
-
-            return {
-              kind: GridCellKind.Text,
-              allowOverlay: true,
-              readonly: true,
-              displayData: String(cellValue),
-              data: cellValue,
-            };
-          }
-        }
-
-        return blankCell;
-      },
-    [columns, router],
-  );
-
-  if (!columns || !rows) {
-    return null;
-  }
 
   const isEmpty = entitiesCount.namespace + entitiesCount.public === 0;
 
@@ -169,14 +101,7 @@ export const EntitiesTab: FunctionComponent = () => {
               height: "50vh",
             }}
           >
-            <Grid
-              showSearch={showSearch}
-              onSearchClose={() => setShowSearch(false)}
-              columns={columns}
-              rows={rows}
-              createGetCellContent={createGetCellContent}
-              customRenderers={[renderTextIconCell]}
-            />
+            <EntitiesTable />
           </Box>
         )}
       </SectionWrapper>
